Add socket URI passing and searching
This commit adds functionality for the tool to either receive or attempt
to identify the correct socket path set up by the service.

A new CLI argument was introduced so the URI can be passed directly. If
the argument isn't used, the tool attempts to use a predefined
environment variable, or other default paths if that is not found
either.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>