Align crates version with parsec-service
For packaging parsec service and the parsec tool to different
distributions, it is ideal to have crate versions aligned so
that multiple versions of the same crate are not included. To
achieve this, the lock file has been updated by running cargo update
generic and also precisely to a particular version for a few crates.

Signed-off-by: Gowtham Suresh Kumar <gowtham.sureshkumar@arm.com>