ci/nightly: Track and test dependencies' 'test-next' branch
parsec-tool depends on several repositories under the parallaxsecond
organization.

Currently, when testing the main branch of parsec-tool, fixed versions
of the latest released crates of those repositories are being used.
A problem may arise when there is a change introduced in the
main/'test-next' branch of those repositories that would break
parsec-tool when updating/incorporating that change. This is currently
not being tested in our CI.
As parsec-tool release is dependent on having published the
dependencies' crates, there are issues that will not be caught until
the publishing of said crates.

* Add a nightly job that uses ci.sh --test-next-branch-tracking to
  test the test-next release branches of parallaxsecond repositories
  and spot issues before release. These job makes use of previously
  defined re-usable actions.

Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>