utils/dependency_cross_matcher.py: Avoid ordering mismatch error
The symmetric difference between versions of parsec and parsec-tool
is computed using sets. The exception list has a specific order, so
when checking whether the version differences between parsec and
parsec-tool are allowed (are in the same as the ones in the
exception list), this may appear to not be allowed.

 * Sort the exception lists and the list obtained from calculating
   the symmetric difference so that the comparison between them can
   be performed without ordering issues.

Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>