parsec-cli-tests.sh: Request expected CSR output format.
OpenSSL 3.2 change the default format for subject and issuer. The
obvious part is that a space is missing but the output also is now using
UTF8.

To keep the format and simplify parsing, openssl can be taught which
output format it should use. This is understood also by older openssl
versions.

Use -nameopt to tell openssl to use UTF8 and a space which around `='
which is what the old parsing expected. Remove the conditional grep
which was introduced to deal with the situation.

Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>